feat(chain): make the chain-provider transport wasm-reachable
Relocate the shared chain-provider constructor (build_chain_provider plus
SharedChainProvider) out of the native-only vertex-swarm-builder into
vertex-chain behind a new `provider` feature, so the wasm SWAP client can
construct a SharedChainProvider for on-chain cashout over a browser fetch
transport.

The alloy-provider transport is target-split in vertex-chain exactly as the
chequebook crate does: native HTTP over reqwest with system TLS on native,
browser fetch (reqwest, TLS via the browser) on wasm32. The provider feature
also pulls the local wallet signer and a single tracing line. The wasm32
getrandom `js` table the k256 signer path needs was already present.

The builder keeps the config and spec resolution in build_node_chain_provider
and calls into the relocated constructor, mapping its ProviderBuildError onto
SwarmNodeError::Chain. The builder's `chain` feature now enables
vertex-chain/provider, and its now-errant direct alloy-provider dependency
(and the unconditional reqwest-native-tls) is removed.
