+++ crates/chain/Cargo.toml
@@ -11,6 +11,17 @@ description = "Chain (Ethereum RPC) configuration, errors, and provider extensio
 [lints]
 workspace = true
 
+[features]
+# Default is the pure, transport-free layer: the `Provider` extension trait,
+# the error taxonomy, the address book. No reqwest, no signer, so the crate
+# stays wasm-safe by default.
+default = []
+# `provider` adds the shared-provider construction seam (`build_chain_provider`
+# and `SharedChainProvider`). It pulls a concrete HTTP transport (target-split:
+# native TLS on native, browser fetch on wasm32) and the local wallet signer, so
+# a launch path never names `alloy-provider` or `alloy-signer-local` directly.
+provider = ["dep:alloy-signer-local", "dep:tracing", "alloy-provider/reqwest"]
+
 [dependencies]
 # nectar
 nectar-contracts = { workspace = true }
@@ -19,9 +30,9 @@ nectar-swarms = { workspace = true }
 # alloy
 alloy-chains = { workspace = true }
 alloy-primitives = { workspace = true }
-# The `Provider` trait alone, with no transport: `default-features = false` drops
-# reqwest and TLS so the extension trait stays wasm-compatible. A concrete
-# transport is selected by the consumer, not here.
+# The `Provider` trait, with no transport by default: `default-features = false`
+# drops reqwest and TLS so the extension trait stays wasm-compatible. The
+# `provider` feature selects a concrete transport (target-split below).
 alloy-provider = { workspace = true }
 alloy-network = { workspace = true }
 alloy-rpc-types-eth = { workspace = true }
@@ -29,15 +40,31 @@ alloy-sol-types = { workspace = true }
 # `alloy-contract` errors flow into the chain error taxonomy so consumers that
 # drive contracts through `CallBuilder` propagate failures with `?`.
 alloy-contract = { workspace = true }
+# `provider` feature: the local wallet signer the construction seam wallet-fills
+# the provider with. Off by default so the transport-free layer stays wasm-safe.
+alloy-signer-local = { workspace = true, optional = true }
 
 # misc
 strum = { workspace = true }
 thiserror = { workspace = true }
+# `provider` feature: a single chain-connected log line in the construction seam.
+tracing = { workspace = true, optional = true }
+
+# Native uses the reqwest transport with native-tls (system OpenSSL). The rustls
+# path drags in a root-certs crate under a license outside the workspace
+# allow-list, so native TLS goes through system OpenSSL.
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+alloy-provider = { workspace = true, features = ["reqwest-native-tls"] }
 
 # alloy's signer stack pulls getrandom 0.2 transitively; on wasm32 its browser
 # backend needs the `js` feature. This direct dep only flips that feature on.
+# The reqwest transport on wasm uses the browser fetch backend, which does TLS
+# through the browser, so no `*-tls` feature is selected.
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = { workspace = true, features = ["js"] }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["rt", "macros"] }
+
+[target.'cfg(target_arch = "wasm32")'.dev-dependencies]
+wasm-bindgen-test = { workspace = true }
